Issue 12024: Merged english i18n resources from RB4, added db diff which reflects changes in RB4 ImageDomainObject, synced ImageDomainObject and related classes with RB4
git-svn-id: https://svn.imcode.com/imcode/imcms/trunk@28209 bd66a97b-2aff-0310-9095-89ca5cabf5a6

+++ src/main/java/com/imcode/imcms/api/ImageHistory.java
@@ -52,7 +52,7 @@ public class ImageHistory {
     private Integer type;
 
     @Column(name = "format", nullable = false)
-    private short format;
+    private int format;
 
     @Column(name = "crop_x1", nullable = false)
     private int cropX1;
@@ -67,7 +67,7 @@ public class ImageHistory {
     private int cropY2;
 
     @Column(name = "rotate_angle", nullable = false)
-    private short rotateAngle;
+    private int rotateAngle;
 
     @Column(name = "gen_file")
     private String generatedFilename;
@@ -279,7 +279,7 @@ public class ImageHistory {
     }
 
     public void setFormat(Format format) {
-        this.format = (short) (format != null ? format.getOrdinal() : 0);
+        this.format = format != null ? format.getOrdinal() : 0;
     }
 
     public CropRegion getCropRegion() {
